Restructures project and renames main packages and app class
