+++ build.gradle
@@ -4,7 +4,7 @@ plugins {
 	id 'java'
 }
 
-group = 'org.launchcode-iftoff'
+group = 'org.launchcode'
 version = '0.0.1-SNAPSHOT'
 sourceCompatibility = '11'
