--- hello world/hello-world/settings.gradle
@@ -1,1 +1,0 @@
-rootProject.name = 'hello-world'
