+++ src/main/java/org/launchcode/mentalhealthcareaccess/MentalHealthcareAccessApplication.java
@@ -1,13 +1,13 @@
-package org.launchcodeiftoff.hello.world;
+package org.launchcode.mentalhealthcareaccess;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class HelloWorldApplication {
+public class MentalHealthcareAccessApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(HelloWorldApplication.class, args);
+		SpringApplication.run(MentalHealthcareAccessApplication.class, args);
 	}
 
 }
